+++ changelog.md
@@ -3,6 +3,7 @@ v?.?.?
 
 - Added a generic "Scientific Calculator", closes #98.
 - Added a dew point calculator based on the Magnus equation, closes #104.
+- Added calculator template search functionality, closes #63.
 - Added scrolling to the calculator selection grid, closes #105.
 - Fixed bug where the MetricPrefix library did not round correctly when no prefix was required (i.e. value is between 1-1000), closes #92.
 - Fixed bug where setting V = 1V, R = 1R gives I = 1E0 in the "Ohm's Law" calculator, closes #89.